requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
ecdsa==0.19.1
fastapi==0.116.1
google-auth==2.40.3
google-genai==1.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kerykeion==4.26.3
numerology==1.8
passlib==1.7.4
platformdirs==4.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyswisseph==2.10.3.2
python-jose==3.5.0
pytz==2024.2
requests==2.32.5
requests-cache==1.2.1
rsa==4.9.1
scour==0.38.2
simple-ascii-tables==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
stripe==12.4.0
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
supabase==2.18.1
python-dotenv==1.1.1
groq==0.4.1
pykakasi==2.2.1
googlemaps==4.10.0
